pyproject.toml
[tool.poetry]
name = "django_starter"
version = "1.0.0"
description = ""
authors = ["Adrian Geuß <adrian@viagis.app>"]

[tool.poetry.dependencies]
python = "^3.7.3"
django = "^3.2"
psycopg2 = "^2.8"
taskipy = "^1"
python-dotenv = "^0.17.0"
djangorestframework = "^3.12.4"
drf-nested-routers = "^0.93.3"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
safety = "*"
flake8 = "^3"
fabric = "^2"
django-debug-toolbar = "^3.2"
shellescape = "^3.8.1"
django-stubs = "^1.8.0"
djangorestframework-stubs = "^1.4.0"

[tool.taskipy.tasks]
django = "python django_starter/manage.py"
lint = "flake8"
security-check = "safety check"
deploy-check = "task django check --deploy"
check = "task lint; task security-check; task deploy-check"
makemessages = "cd django && ./manage.py makemessages -l en -l de"  # add languages as required
compilemessages = "task django compilemessages --ignore .venv"
build = "task compilemessages"  # Add additional tasks with '&& task'
pre_deploy = "task check; task build"
deploy = "python django_starter/deploy.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
